refactor: Remove logout functionality from FormList component
This commit removes the logout functionality from the FormList component. The handleLogout function and related state variables have been removed, as well as the logout button in the component's JSX. This change aligns with the recent user commits that focused on updating the Home page layout and does not affect the functionality of the FormList component.