refactor: Update FormList component to include loading screen color customization
This commit updates the FormList component to include the ability to customize the color of the loading screen based on the current theme. It adds a new optional "color" prop to the LoadingScreen component and uses the "useTheme" hook from "next-themes" to determine the current theme. The color is set to white when the theme is dark and to #707070 when the theme is light. This change improves the visual consistency of the loading screen with the overall theme of the application.