refactor: Update DeleteFormPopUp and FormItem components
This commit updates the DeleteFormPopUp and FormItem components to improve consistency in the application. In the DeleteFormPopUp component, the margin-top (mt-3) class is used instead of the margin (m-3) class to ensure consistent spacing. In the FormItem component, the background color of the container div is removed to allow for more flexibility in styling. These changes align with recent user commits that focused on updating the layout and styling of various components.