refactor: Add resetPassword function to AuthContext
The resetPassword function has been added to the AuthContext to handle the process of resetting a user's password. This function makes a POST request to the "/request-reset-password" endpoint with the provided user data. If the request is successful, the user's auth data is updated and a success message is stored in sessionStorage. This change aligns with recent user commits that focused on improving the functionality and user experience of the authentication process.